+++ weblate/urls.py
@@ -520,6 +520,11 @@ real_patterns = [
         name='component_progress',
     ),
     url(
+        r'^progress/' + COMPONENT + 'terminate/$',
+        weblate.trans.views.manage.component_progress_terminate,
+        name='component_progress_terminate',
+    ),
+    url(
         r'^js/progress/' + COMPONENT + '$',
         weblate.trans.views.manage.component_progress_js,
         name='component_progress_js',