+++ docs/changes.rst
@@ -7,6 +7,7 @@ weblate 3.7
 Released on ? 2019.
 
 * Added separate Celery queue for notifications.
+* Use consistent look with application for API browsing.
 
 weblate 3.6.1
 -------------